feat(icons): rename module icons Rocket→Onboarding and ItManagement→Software
- Rename modules/Rocket to modules/Onboarding (matches Figma 'Onboarding' node)
- Remove modules/ItManagement duplicate (identical SVG to Software, which already exists)
- Update F0AvatarModule/modules.ts: get_started→Onboarding, it_management→Software
- Apply same changes to react-native package
