--- packages/react-native/src/icons/modules/ItManagement.tsx
@@ -1,19 +1,0 @@
-import { Ref, forwardRef } from "react"
-import Svg, { Path } from "react-native-svg"
-import type { SvgProps } from "react-native-svg"
-const SvgItManagement = (props: SvgProps, ref: Ref<Svg>) => (
-  <Svg
-    fill="none"
-    viewBox="0 0 24 24"
-    className={props.className}
-    ref={ref}
-    {...props}
-  >
-    <Path
-      fill="currentColor"
-      d="M3 7.5a3.6 3.6 0 0 1 3.6-3.6h10.8A3.6 3.6 0 0 1 21 7.5v5.4a3.6 3.6 0 0 1-3.6 3.6h-1.8v1.8h.9a.9.9 0 1 1 0 1.8h-9a.9.9 0 0 1 0-1.8h.9v-1.8H6.6A3.6 3.6 0 0 1 3 12.9zm3.6 7.2h10.8a1.8 1.8 0 0 0 1.8-1.8V7.5a1.8 1.8 0 0 0-1.8-1.8H6.6a1.8 1.8 0 0 0-1.8 1.8v5.4a1.8 1.8 0 0 0 1.8 1.8"
-    />
-  </Svg>
-)
-const ForwardRef = forwardRef(SvgItManagement)
-export default ForwardRef
