+++ packages/react-native/src/icons/modules/index.ts
@@ -1,41 +1,41 @@
-export { default as Benefits } from "./Benefits"
-export { default as Calendar } from "./Calendar"
-export { default as Cards } from "./Cards"
-export { default as ClockIn } from "./ClockIn"
-export { default as Complaints } from "./Complaints"
-export { default as Discover } from "./Discover"
-export { default as Documents } from "./Documents"
-export { default as Engagement } from "./Engagement"
-export { default as Finance } from "./Finance"
-export { default as Goals } from "./Goals"
-export { default as Home } from "./Home"
-export { default as Hub } from "./Hub"
-export { default as ItManagement } from "./ItManagement"
-export { default as Inbox } from "./Inbox"
-export { default as Kudos } from "./Kudos"
-export { default as Marketplace } from "./Marketplace"
-export { default as Meetings } from "./Meetings"
-export { default as MyDocuments } from "./MyDocuments"
-export { default as Organization } from "./Organization"
-export { default as Overviews } from "./Overviews"
-export { default as Pages } from "./Pages"
-export { default as Payroll } from "./Payroll"
-export { default as Performance } from "./Performance"
-export { default as Profile } from "./Profile"
-export { default as Projects } from "./Projects"
-export { default as Recruitment } from "./Recruitment"
-export { default as Reports } from "./Reports"
-export { default as SalaryAdvance } from "./SalaryAdvance"
-export { default as Sales } from "./Sales"
-export { default as Settings } from "./Settings"
-export { default as Shifts } from "./Shifts"
-export { default as Social } from "./Social"
-export { default as Software } from "./Software"
-export { default as Spaces } from "./Spaces"
-export { default as Spending } from "./Spending"
-export { default as Tasks } from "./Tasks"
-export { default as TimeOff } from "./TimeOff"
-export { default as TimeTracking } from "./TimeTracking"
-export { default as Trainings } from "./Trainings"
-export { default as Treasury } from "./Treasury"
-export { default as Workflows } from "./Workflows"
+export { default as Benefits } from "./Benefits";
+export { default as Calendar } from "./Calendar";
+export { default as Cards } from "./Cards";
+export { default as ClockIn } from "./ClockIn";
+export { default as Complaints } from "./Complaints";
+export { default as Discover } from "./Discover";
+export { default as Documents } from "./Documents";
+export { default as Engagement } from "./Engagement";
+export { default as Finance } from "./Finance";
+export { default as Goals } from "./Goals";
+export { default as Home } from "./Home";
+export { default as Hub } from "./Hub";
+export { default as Onboarding } from "./Onboarding";
+export { default as Inbox } from "./Inbox";
+export { default as Kudos } from "./Kudos";
+export { default as Marketplace } from "./Marketplace";
+export { default as Meetings } from "./Meetings";
+export { default as MyDocuments } from "./MyDocuments";
+export { default as Organization } from "./Organization";
+export { default as Overviews } from "./Overviews";
+export { default as Pages } from "./Pages";
+export { default as Payroll } from "./Payroll";
+export { default as Performance } from "./Performance";
+export { default as Profile } from "./Profile";
+export { default as Projects } from "./Projects";
+export { default as Recruitment } from "./Recruitment";
+export { default as Reports } from "./Reports";
+export { default as SalaryAdvance } from "./SalaryAdvance";
+export { default as Sales } from "./Sales";
+export { default as Settings } from "./Settings";
+export { default as Shifts } from "./Shifts";
+export { default as Social } from "./Social";
+export { default as Software } from "./Software";
+export { default as Spaces } from "./Spaces";
+export { default as Spending } from "./Spending";
+export { default as Tasks } from "./Tasks";
+export { default as TimeOff } from "./TimeOff";
+export { default as TimeTracking } from "./TimeTracking";
+export { default as Trainings } from "./Trainings";
+export { default as Treasury } from "./Treasury";
+export { default as Workflows } from "./Workflows";
