+++ packages/react/src/components/avatars/F0AvatarModule/modules.ts
@@ -32,10 +32,10 @@ export const modules = {
   "finance-treasury": ModuleIcons.Treasury,
   "finance-workspace": ModuleIcons.Reports,
   goals: ModuleIcons.Goals,
-  get_started: ModuleIcons.Rocket,
+  get_started: ModuleIcons.Onboarding,
   home: ModuleIcons.Home,
   hub: ModuleIcons.Hub,
-  it_management: ModuleIcons.ItManagement,
+  it_management: ModuleIcons.Software,
   kudos: ModuleIcons.Kudos,
   lms: ModuleIcons.LMS,
   meetings: ModuleIcons.Meetings,
