+++ packages/react/src/icons/modules/index.ts
@@ -12,7 +12,7 @@ export { default as Goals } from "./Goals"
 export { default as Home } from "./Home"
 export { default as Hub } from "./Hub"
 export { default as Inbox } from "./Inbox"
-export { default as ItManagement } from "./ItManagement"
+export { default as Onboarding } from "./Onboarding"
 export { default as Kudos } from "./Kudos"
 export { default as LMS } from "./LMS"
 export { default as Marketplace } from "./Marketplace"
@@ -27,7 +27,6 @@ export { default as Profile } from "./Profile"
 export { default as Projects } from "./Projects"
 export { default as Recruitment } from "./Recruitment"
 export { default as Reports } from "./Reports"
-export { default as Rocket } from "./Rocket"
 export { default as SalaryAdvance } from "./SalaryAdvance"
 export { default as Sales } from "./Sales"
 export { default as Settings } from "./Settings"
